organises Postgres related code into the database module
