+++ dashboard/src/t5gweb/database/__init__.py
@@ -1,0 +1,51 @@
+"""Database module for t5gweb
+
+This module provides database functionality including:
+- Database models (Case, Comment, JiraCard, JiraComment)
+- Session management (engine, SessionLocal, Base)
+- Database operations (load_cases_postgres, load_jira_cards_postgres)
+- Utility functions (get_db, create_postgres_tables)
+"""
+
+# Import session management components
+from .session import (
+    Base,
+    engine,
+    SessionLocal,
+    create_postgres_tables,
+    get_db,
+)
+
+# Import database models
+from .models import (
+    Case,
+    Comment,
+    JiraCard,
+    JiraComment,
+)
+
+# Import database operations
+from .operations import (
+    load_cases_postgres,
+    load_jira_cards_postgres,
+)
+
+# Export all components that were previously available from postgres_db.py
+__all__ = [
+    # Session management
+    "Base",
+    "engine", 
+    "SessionLocal",
+    "create_postgres_tables",
+    "get_db",
+    
+    # Models
+    "Case",
+    "Comment", 
+    "JiraCard",
+    "JiraComment",
+    
+    # Operations
+    "load_cases_postgres",
+    "load_jira_cards_postgres",
+]
